Modulo Log terminado, Vistas de inicio de sesion terminadas, Responsive terminado, falta llenar información correcta en base de datos y pruebas finales

+++ Vista/Administrador/actualizarInfo.php
@@ -3,32 +3,36 @@ $infoLog = "";
 $administradorAux;
 $admAux = new Administrador($_SESSION["id"]);
 $admAux->traerInfo();
-$band = false;
+$band = 0;
 if (isset($_POST["Actualizar"])) {
     if ($_FILES["imagen"]["name"] != "") {
-        $rutaLocal = $_FILES["imagen"]["tmp_name"];
-        $tipo = $_FILES["imagen"]["type"];
-        $tiempo = new DateTime();
-        $rutaRemota = "Vista/Img/Users/" . $tiempo->getTimestamp() . (($tipo == "image/png") ? ".png" : ".jpg");
-        copy($rutaLocal, $rutaRemota);
-        $administradorAux = new Administrador($_SESSION["id"]);
-        $administradorAux->traerInfo();
+        if ($_FILES["imagen"]["type"] == "image/png" or $_FILES["imagen"]["type"] == "image/jpeg") {
+            $rutaLocal = $_FILES["imagen"]["tmp_name"];
+            $tipo = $_FILES["imagen"]["type"];
+            $tiempo = new DateTime();
+            $rutaRemota = "Vista/Img/Users/" . $tiempo->getTimestamp() . (($tipo == "image/png") ? ".png" : ".jpg");
+            copy($rutaLocal, $rutaRemota);
+            $administradorAux = new Administrador($_SESSION["id"]);
+            $administradorAux->traerInfo();
 
-        if ($administradorAux->getFoto() != "") {
-            unlink($administradorAux->getFoto());
+            if ($administradorAux->getFoto() != "") {
+                unlink($administradorAux->getFoto());
+            }
+            $administrador = new Administrador($_SESSION["id"], $_POST["Nombre"], $_POST["Correo"], "", "", $rutaRemota);
+            $administrador->actualizarInfo();
+            $band = 1;
         }
-        $administrador = new Administrador($_SESSION["id"], $_POST["Nombre"], $_POST["Correo"], "", "", $rutaRemota);
-        $administrador->actualizarInfo();
-        $band = true;
     } else if ($admAux->getNombre() != $_POST["Nombre"] or $admAux->getCorreo() != $_POST["Correo"]) { /*Se valida que haya cambiado algun campo*/
         $administradorAux = new Administrador($_SESSION["id"]);
         $administradorAux->traerInfo();
         $administrador = new Administrador($_SESSION["id"], $_POST["Nombre"], $_POST["Correo"], "", "", $administradorAux->getFoto());
         $administrador->actualizarInfo();
-        $band = true;
+        $band = 1;
+    }else{
+        $band=2;
     }
-    if ($band == true) {
-        $infoLog = "Actor:Administrador-Nombre:" . $administradorAux->getNombre() . "-Correo:" . $administradorAux->getCorreo(). "-id:" . $administradorAux->getIdAdministrador();
+    if ($band == 1) {
+        $infoLog = "Actor:Administrador-Nombre:" . $administradorAux->getNombre() . "-Correo:" . $administradorAux->getCorreo() . "-id:" . $administradorAux->getIdAdministrador();
         date_default_timezone_set('America/Bogota');
         $date = date('Y-m-d');
         $hora = date('H:i:s');
@@ -53,10 +57,10 @@ if (isset($_POST["Actualizar"])) {
                 </div>
                 <div class="card-body">
                     <div class="row p-3">
-                        <div class="col-3">
+                        <div class="col-xl-3 col-lg-3 col-md-12 col-sm-12 col-12">
                             <img src="<?php echo ($administrador->getFoto() != "" ? $administrador->getFoto() : "https://upload.wikimedia.org/wikipedia/commons/e/e4/Elliot_Grieveson.png") ?>" width="100%" class="img-thumbnail">
                         </div>
-                        <div class="col-9">
+                        <div class="col-xl-9 col-lg-9 col-md-12 col-sm-12 col-12">
                             <form action="index.php?pid= <?php echo base64_encode("Vista/Administrador/actualizarInfo.php") ?>" method="POST" enctype="multipart/form-data">
                                 <div class="form-group">
                                     <label for="Nombre">Nombres</label>
@@ -88,13 +92,13 @@ if (isset($_POST["Actualizar"])) {
     </div>
 </div>
 
-<?php if ($band == true and isset($_POST["Actualizar"])) {
+<?php if ($band == 1) {
     echo "<div class='modal fade' id='mostrarmodal' tabindex='-1' role='dialog' aria-labelledby='exampleModalLabel' aria-hidden='true'>";
     echo "<div class='modal-dialog '>";
     echo "<div class='modal-content '>";
     echo "<div class='alert alert-success m-0 d-flex justify-content-between'>";
     echo "<div></div>";
-    echo "<h5 class='modal-title flex-grow-1' id='exampleModalLabel'>La información se ha actualizado correctamente</h5>";
+    echo "<h5 class='modal-title' id='exampleModalLabel'>La información se ha actualizado correctamente</h5>";
     echo "<button type='button' class='close' data-dismiss='modal' aria-label='Close'>";
     echo "<span aria-hidden='true'>&times;</span>";
     echo "</button>";
@@ -102,13 +106,27 @@ if (isset($_POST["Actualizar"])) {
     echo "</div>";
     echo "</div>";
     echo "</div>";
-} else if ($band == false and isset($_POST["Actualizar"])) {
+} else if (isset($_POST["Actualizar"]) and $band == 0) {
+    echo "<div class='modal fade' id='mostrarmodal' tabindex='-1' role='dialog' aria-labelledby='exampleModalLabel' aria-hidden='true'>";
+    echo "<div class='modal-dialog '>";
+    echo "<div class='modal-content '>";
+    echo "<div class='alert alert-danger m-0 d-flex justify-content-between'>";
+    echo "<div></div>";
+    echo "<h5 class='modal-title' id='exampleModalLabel'>Error en el tipo de archivo</h5>";
+    echo "<button type='button' class='close' data-dismiss='modal' aria-label='Close'>";
+    echo "<span aria-hidden='true'>&times;</span>";
+    echo "</button>";
+    echo "</div>";
+    echo "</div>";
+    echo "</div>";
+    echo "</div>";
+} else if ($band == 2) {
     echo "<div class='modal fade' id='mostrarmodal' tabindex='-1' role='dialog' aria-labelledby='exampleModalLabel' aria-hidden='true'>";
     echo "<div class='modal-dialog '>";
     echo "<div class='modal-content '>";
     echo "<div class='alert alert-warning m-0 d-flex justify-content-between'>";
     echo "<div></div>";
-    echo "<h5 class='modal-title' id='exampleModalLabel'>No se ha modificado ningún dato</h5>";
+    echo "<h5 class='modal-title' id='exampleModalLabel'>No se ha cambiado ningun dato</h5>";
     echo "<button type='button' class='close' data-dismiss='modal' aria-label='Close'>";
     echo "<span aria-hidden='true'>&times;</span>";
     echo "</button>";
