+++ Vista/Producto/editarProducto.php
@@ -1,45 +1,48 @@
 <?php
 $infoLog = "";
 $productoAux;
-$band = false;
+$band = 0;
 $proAux;
-echo $_GET["idProducto"];
 $proAux = new Producto($_GET["idProducto"]);
 $proAux->traerInfo();
 
 if (isset($_POST["Actualizar"])) {
     if ($_FILES["imagen"]["name"] != "") {
-        $rutaLocal = $_FILES["imagen"]["tmp_name"];
-        $tipo = $_FILES["imagen"]["type"];
-        $tiempo = new DateTime();
-        $rutaRemota = "Vista/Img/Users/" . $tiempo->getTimestamp() . (($tipo == "image/png") ? ".png" : ".jpg");
-        copy($rutaLocal, $rutaRemota);
-        $productoAux = new Producto($_POST["idProducto"]);
-        $productoAux->traerInfo();
+        if ($_FILES["imagen"]["type"] == "image/png" or $_FILES["imagen"]["type"] == "image/jpeg") {
+            $rutaLocal = $_FILES["imagen"]["tmp_name"];
+            $tipo = $_FILES["imagen"]["type"];
+            $tiempo = new DateTime();
+            $rutaRemota = "Vista/Img/imgProductos/" . $tiempo->getTimestamp() . (($tipo == "image/png") ? ".jpeg" : ".jpg");
+            copy($rutaLocal, $rutaRemota);
+            $productoAux = new Producto($_POST["idProducto"]);
+            $productoAux->traerInfo();
 
-        if ($productoAux->getImagen() != "") {
-            unlink($productoAux->getImagen());
+            if ($productoAux->getImagen() != "") {
+                unlink($productoAux->getImagen());
+            }
+            $producto = new Producto($_POST["idProducto"], $_POST["Nombre"], $_POST["Cantidad"], $_POST["Precio"], $rutaRemota);
+            $producto->actualizarProducto();
+            $band = 1;
         }
-        $producto = new Producto($_POST["idProducto"], $_POST["Nombre"], $_POST["Cantidad"], $_POST["Precio"], $rutaRemota);
-        $producto->actualizarProducto();
-        $band = true;
     } else if ($proAux->getNombre() != $_POST["Nombre"] or $proAux->getCantidad() != $_POST["Cantidad"] or $proAux->getPrecio() != $_POST["Precio"]) { /*Se valida que haya cambiado algun campo*/
         $productoAux = new Producto($_POST["idProducto"]);
         $productoAux->traerInfo();
         $producto = new Producto($_POST["idProducto"], $_POST["Nombre"], $_POST["Cantidad"], $_POST["Precio"], $productoAux->getImagen());
         $producto->actualizarProducto();
-        $band = true;
+        $band = 1;
+    }else{
+        $band=2;
     }
-    if ($band == true) {
-        $infoLog = "id:".$productoAux->getidProducto()."-Nombre:" . $productoAux->getNombre() . "-Cantidad:" . $productoAux->getCantidad() . "-Precio:" . $productoAux->getPrecio();
+    if ($band == 1) {
+        $infoLog = "id:" . $productoAux->getidProducto() . "-Nombre:" . $productoAux->getNombre() . "-Cantidad:" . $productoAux->getCantidad() . "-Precio:" . $productoAux->getPrecio();
         date_default_timezone_set('America/Bogota');
         $date = date('Y-m-d');
         $hora = date('H:i:s');
         $log = new Log("", 'Editar Producto', $infoLog, $date, $hora, "Administrador", $_SESSION["id"]);
         $log->insertarLog();
-    }else{
+    } else {
         $producto = new Producto($_GET["idProducto"]);
-        $producto->traerInfo(); 
+        $producto->traerInfo();
     }
 } else {
     $producto = new Producto($_GET["idProducto"]);
@@ -57,10 +60,10 @@ if (isset($_POST["Actualizar"])) {
                 </div>
                 <div class="card-body">
                     <div class="row p-3">
-                        <div class="col-3">
+                        <div class="col-xl-3 col-lg-3 col-md-12 col-sm-12 col-12">
                             <img src="<?php echo ($producto->getImagen() != "" ? $producto->getImagen() : "https://icons.iconarchive.com/icons/xaml-icon-studio/agriculture/128/Fruits-Vegetables-icon.png") ?>" width="100%" class="img-thumbnail">
                         </div>
-                        <div class="col-9">
+                        <div class="col-xl-9 col-lg-9 col-md-12 col-sm-12 col-12">
                             <form action="index.php?pid= <?php echo base64_encode("Vista/Producto/editarProducto.php") ?>&idProducto=<?php echo $_GET["idProducto"] ?>" method="POST" enctype="multipart/form-data">
                                 <div class="form-group">
                                     <label for="idProducto">ID Producto</label>
@@ -99,13 +102,13 @@ if (isset($_POST["Actualizar"])) {
     </div>
 </div>
 
-<?php if ($band == true and isset($_POST["Actualizar"])) {
+<?php if ($band == 1){
     echo "<div class='modal fade' id='mostrarmodal' tabindex='-1' role='dialog' aria-labelledby='exampleModalLabel' aria-hidden='true'>";
     echo "<div class='modal-dialog '>";
     echo "<div class='modal-content '>";
     echo "<div class='alert alert-success m-0 d-flex justify-content-between'>";
     echo "<div></div>";
-    echo "<h5 class='modal-title flex-grow-1' id='exampleModalLabel'>La información se ha actualizado correctamente</h5>";
+    echo "<h5 class='modal-title' id='exampleModalLabel'>La información se ha actualizado correctamente</h5>";
     echo "<button type='button' class='close' data-dismiss='modal' aria-label='Close'>";
     echo "<span aria-hidden='true'>&times;</span>";
     echo "</button>";
@@ -113,13 +116,27 @@ if (isset($_POST["Actualizar"])) {
     echo "</div>";
     echo "</div>";
     echo "</div>";
-} else if ($band == false and isset($_POST["Actualizar"])) {
+} else if (isset($_POST["Actualizar"]) and $band == 0) {
+    echo "<div class='modal fade' id='mostrarmodal' tabindex='-1' role='dialog' aria-labelledby='exampleModalLabel' aria-hidden='true'>";
+    echo "<div class='modal-dialog '>";
+    echo "<div class='modal-content '>";
+    echo "<div class='alert alert-danger m-0 d-flex justify-content-between'>";
+    echo "<div></div>";
+    echo "<h5 class='modal-title' id='exampleModalLabel'>Error en el tipo de archivo</h5>";
+    echo "<button type='button' class='close' data-dismiss='modal' aria-label='Close'>";
+    echo "<span aria-hidden='true'>&times;</span>";
+    echo "</button>";
+    echo "</div>";
+    echo "</div>";
+    echo "</div>";
+    echo "</div>";
+}else if($band== 2){
     echo "<div class='modal fade' id='mostrarmodal' tabindex='-1' role='dialog' aria-labelledby='exampleModalLabel' aria-hidden='true'>";
     echo "<div class='modal-dialog '>";
     echo "<div class='modal-content '>";
     echo "<div class='alert alert-warning m-0 d-flex justify-content-between'>";
     echo "<div></div>";
-    echo "<h5 class='modal-title' id='exampleModalLabel'>No se ha modificado ningún dato</h5>";
+    echo "<h5 class='modal-title' id='exampleModalLabel'>No se ha cambiado ningun dato</h5>";
     echo "<button type='button' class='close' data-dismiss='modal' aria-label='Close'>";
     echo "<span aria-hidden='true'>&times;</span>";
     echo "</button>";
