+++ Vista/Proveedor/editarInfo.php
@@ -3,32 +3,37 @@ $infoLog = "";
 $proveedorAux;
 $proAux = new Proveedor($_GET["idProveedor"]);
 $proAux->traerInfo();
-$band = false;
+$band = 0;
 if (isset($_POST["Editar"])) {
     if ($_FILES["imagen"]["name"] != "") {
-        $rutaLocal = $_FILES["imagen"]["tmp_name"];
-        $tipo = $_FILES["imagen"]["type"];
-        $tiempo = new DateTime();
-        $rutaRemota = "Vista/Img/Users/" . $tiempo->getTimestamp() . (($tipo == "image/png") ? ".png" : ".jpg");
-        copy($rutaLocal, $rutaRemota);
-        $proveedorAux = new Proveedor($_GET["idProveedor"]);
-        $proveedorAux->traerInfo();
+        if ($_FILES["imagen"]["type"] == "image/png" or $_FILES["imagen"]["type"] == "image/jpeg") {
+            $rutaLocal = $_FILES["imagen"]["tmp_name"];
+            $tipo = $_FILES["imagen"]["type"];
+            echo $tipo;
+            $tiempo = new DateTime();
+            $rutaRemota = "Vista/Img/Users/" . $tiempo->getTimestamp() . (($tipo == "image/png") ? ".png" : ".jpeg");
+            copy($rutaLocal, $rutaRemota);
+            $proveedorAux = new Proveedor($_GET["idProveedor"]);
+            $proveedorAux->traerInfo();
 
-        if ($proveedorAux->getFoto() != "") {
-            unlink($proveedorAux->getFoto());
+            if ($proveedorAux->getFoto() != "") {
+                unlink($proveedorAux->getFoto());
+            }
+            $proveedor = new Proveedor($_GET["idProveedor"], $_POST["Nombre"], $_POST["Correo"], "", "", $rutaRemota);
+            $proveedor->actualizarInfo();
+            $band = 1;
         }
-        $proveedor = new Proveedor($_GET["idProveedor"], $_POST["Nombre"], $_POST["Correo"], "", "", $rutaRemota);
-        $proveedor->actualizarInfo();
-        $band = true;
     } else if ($proAux->getNombre() != $_POST["Nombre"] or $proAux->getCorreo() != $_POST["Correo"]) { /*Se valida que haya cambiado algun campo*/
         $proveedorAux = new Proveedor($_GET["idProveedor"]);
         $proveedorAux->traerInfo();
         $proveedor = new Proveedor($_GET["idProveedor"], $_POST["Nombre"], $_POST["Correo"], "", "", $proveedorAux->getFoto());
         $proveedor->actualizarInfo();
-        $band = true;
+        $band = 1;
+    }else{
+        $band=2;
     }
-    if ($band == true) {
-        $infoLog = "Actor:Proveedor-Nombre:" . $proveedorAux->getNombre() . "-Correo:" . $proveedorAux->getCorreo(). "-id:" . $proveedorAux->getIdProveedor();
+    if ($band == 1) {
+        $infoLog = "Actor:Proveedor-Nombre:" . $proveedorAux->getNombre() . "-Correo:" . $proveedorAux->getCorreo() . "-id:" . $proveedorAux->getIdProveedor();
         date_default_timezone_set('America/Bogota');
         $date = date('Y-m-d');
         $hora = date('H:i:s');
@@ -53,11 +58,11 @@ if (isset($_POST["Editar"])) {
                 </div>
                 <div class="card-body">
                     <div class="row p-3">
-                        <div class="col-3">
+                        <div class="col-xl-3 col-lg-3 col-md-12 col-sm-12 col-12">
                             <img src="<?php echo ($proveedor->getFoto() != "" ? $proveedor->getFoto() : "https://upload.wikimedia.org/wikipedia/commons/e/e4/Elliot_Grieveson.png") ?>" width="100%" class="img-thumbnail">
                         </div>
-                        <div class="col-9">
-                            <form action="index.php?pid= <?php echo base64_encode("Vista/Proveedor/editarInfo.php") ?>&idProveedor=<?php echo $_GET["idProveedor"]?>" method="POST" enctype="multipart/form-data">
+                        <div class="col-xl-9 col-lg-9 col-md-12 col-sm-12 col-12">
+                            <form action="index.php?pid= <?php echo base64_encode("Vista/Proveedor/editarInfo.php") ?>&idProveedor=<?php echo $_GET["idProveedor"] ?>" method="POST" enctype="multipart/form-data">
                                 <div class="form-group">
                                     <label for="Nombre">Nombres</label>
                                     <input type="text" class="form-control" name="Nombre" id="Nombre" value="<?php echo $proveedor->getNombre() ?>">
@@ -88,12 +93,41 @@ if (isset($_POST["Editar"])) {
     </div>
 </div>
 
-<?php if (isset($_POST["Editar"])) {
+<?php if ($band == 1){
     echo "<div class='modal fade' id='mostrarmodal' tabindex='-1' role='dialog' aria-labelledby='exampleModalLabel' aria-hidden='true'>";
     echo "<div class='modal-dialog '>";
     echo "<div class='modal-content '>";
-    echo "<div class='alert alert-success m-0 d-flex'>";
-    echo "<h5 class='modal-title flex-grow-1' id='exampleModalLabel'>La información se ha actualizado correctamente</h5>";
+    echo "<div class='alert alert-success m-0 d-flex justify-content-between'>";
+    echo "<div></div>";
+    echo "<h5 class='modal-title' id='exampleModalLabel'>La información se ha actualizado correctamente</h5>";
+    echo "<button type='button' class='close' data-dismiss='modal' aria-label='Close'>";
+    echo "<span aria-hidden='true'>&times;</span>";
+    echo "</button>";
+    echo "</div>";
+    echo "</div>";
+    echo "</div>";
+    echo "</div>";
+} else if (isset($_POST["Editar"]) and $band == 0) {
+    echo "<div class='modal fade' id='mostrarmodal' tabindex='-1' role='dialog' aria-labelledby='exampleModalLabel' aria-hidden='true'>";
+    echo "<div class='modal-dialog '>";
+    echo "<div class='modal-content '>";
+    echo "<div class='alert alert-danger m-0 d-flex justify-content-between'>";
+    echo "<div></div>";
+    echo "<h5 class='modal-title' id='exampleModalLabel'>Error en el tipo de archivo</h5>";
+    echo "<button type='button' class='close' data-dismiss='modal' aria-label='Close'>";
+    echo "<span aria-hidden='true'>&times;</span>";
+    echo "</button>";
+    echo "</div>";
+    echo "</div>";
+    echo "</div>";
+    echo "</div>";
+}else if($band== 2){
+    echo "<div class='modal fade' id='mostrarmodal' tabindex='-1' role='dialog' aria-labelledby='exampleModalLabel' aria-hidden='true'>";
+    echo "<div class='modal-dialog '>";
+    echo "<div class='modal-content '>";
+    echo "<div class='alert alert-warning m-0 d-flex justify-content-between'>";
+    echo "<div></div>";
+    echo "<h5 class='modal-title' id='exampleModalLabel'>No se ha cambiado ningun dato</h5>";
     echo "<button type='button' class='close' data-dismiss='modal' aria-label='Close'>";
     echo "<span aria-hidden='true'>&times;</span>";
     echo "</button>";
